refactor: extract some components
